Mejoro un poco la configuracion del juego

+++ server_src/Entities/PlayerNet.cpp
@@ -77,7 +77,7 @@ float PlayerNet::getIntelligence() {
 }
 
 float PlayerNet::getStrength() {
-  int baseStrength =
+  float baseStrength =
       playerRace->getStrength() * playerClass->getStrengthFactor();
   return equation::getCurrentLevelStrength(baseStrength, level);
 }
@@ -231,7 +231,7 @@ void PlayerNet::updateMaxMana() {
                               playerRace->getManaFactor(), level);
 }
 
-int PlayerNet::getConstitution() {
+float PlayerNet::getConstitution() {
   return playerClass->getConstitutionFactor() * playerRace->getConstitution();
 }
