+++ server_src/headers/Equations.h
@@ -36,7 +36,7 @@ bool isCritical();
 
 bool dodgeAttack(int agility);
 
-int getCurrentLevelStrength(int baseStrength, int level);
+float getCurrentLevelStrength(float baseStrength, int level);
 }  // namespace equation
 
 #endif  // EQUATIONS_H
